+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The *alited* provides the rich and comprehensive set of features, some are uniqu
 # LAST CHANGES
 
 
-Version `v1.8.7.5 (28 Dec'24)`
+Version `v1.8.7.6 (30 Dec'24)`
 
   - BUGFIX: "Save as" with duplicate tab name, e.g. FNAME, should be FNAME(2)
   - BUGFIX: creating a unit tree item at "proc A{a}" (not "proc A {a}")
@@ -16,6 +16,7 @@ Version `v1.8.7.5 (28 Dec'24)`
   - CHANGE: main::UpdateUnitTree before checking Tcl, for assurance
   - CHANGE: "Copy Declaration" in units/favorites menu
   - CHANGE: clearance
+  - CHANGE: packages: apave/e_menu 4.4.11, bartabs 1.6.10
 
 
 Version `v1.8.7 (9 Oct'24)`